requirements.txt
# Database drivers
psycopg2-binary==2.9.7
PyMySQL==1.1.0
asyncpg==0.28.0

# CLI and UI
click==8.1.7
rich==13.5.2
tqdm==4.66.1

# Data processing
pandas==2.1.0
numpy==1.24.3
scipy==1.11.2

# Configuration and serialization
pyyaml==6.0.1
toml==0.10.2

# System monitoring
psutil==5.9.5

# Web reporting
jinja2==3.1.2
plotly==5.15.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# Time handling
pytz==2023.3
